Restructure jem for CEM compatibility
The jem files provided in this repository are designed for compatibility
with Custom Entity Models, and not Optifine, which uses a different
directory structure.

By making these files available at the correct path, this opens up the
potential to contribute this back to the upstream FreshAnimations
project, and maintain compatibility with both Optifine and CEM in the
same resource pack.